Read signatures from metadata and keep them whole
A signature model decoded from the module's own metadata keeps custom
modifiers, array shapes, function pointer signatures, pinning, and generic
parameters by position; reflection types serve the stack simulator only. The
renderer now quotes member names, keeps field modifiers, and spells function
pointer types. Under instance explicit the receiver is popped once, and
float32(0x...) literals keep their bits through both emitters.

+++ src/IlRepl.Engine/Engine/InstructionParser.cs
@@ -252,7 +252,7 @@ public static class InstructionParser
                 return new Instruction { Op = op, Text = text, Kind = OperandKind.Int64, Operand = LiteralParser.ParseInteger(operandText, opName) };
 
             case OperandType.ShortInlineR:
-                return new Instruction { Op = op, Text = text, Kind = OperandKind.Single, Operand = (float)LiteralParser.ParseFloat(operandText, opName) };
+                return new Instruction { Op = op, Text = text, Kind = OperandKind.Single, Operand = LiteralParser.ParseFloat32(operandText, opName) };
 
             case OperandType.InlineR:
                 return new Instruction { Op = op, Text = text, Kind = OperandKind.Double, Operand = LiteralParser.ParseFloat(operandText, opName) };
